Cleaned up Javascript File. Added Modal changes, Extra content for one section added
Modal is more accurate in size and easier to read.

Hiding and showing taskbar function was made so that you can state showing or hiding in parameters

hiding and showing extra content on click now embeded for one list item.

Added a function which auto indents content. This may be changed in future so no inline styles are applied through javascript.